check file exists before attach new (uuid by body file)

+++ allure/common.py
@@ -206,22 +206,16 @@ class AllureImpl(object):
 
         :arg body: str or unicode with contents. str is written as-is in byte stream, unicode is written as utf-8 (what do you expect else?)
         """
-        with self._attachfile("%s-attachment.%s" % (uuid.uuid4(), attach_type.extension)) as f:
+        filename = "%s-attachment.%s" % (uuid.uuid3(uuid.NAMESPACE_DNS, body), attach_type.extension)
+        reportpath = os.path.join(self.logdir, filename)
+        if os.path.exists(reportpath):
+            return os.path.basename(filename)
+        with open(reportpath, 'wb') as f:
             if isinstance(body, text_type):
                 f.write(body.encode('utf-8'))
             else:
                 f.write(body)
             return os.path.basename(f.name)
-
-    @contextmanager
-    def _attachfile(self, filename):
-        """
-        Yields open file object in the report directory with given name
-        """
-        reportpath = os.path.join(self.logdir, filename)
-
-        with open(reportpath, 'wb') as f:
-            yield f
 
     @contextmanager
     def _reportfile(self, filename):
